Secrets: Fixed pet HP resetting if the pet is above 32k when zoning.
Secrets: Fixed player HP rollover bug (~495k hitpoints, rarely encountered) with AAs by temporarily casting it as a float.

Secrets: Fixed damage on weapons above 255. They only display on SoF+ clients, and Titanium will simply show a cosmetic issue (an incorrect value). You can set an item to any value up to the max for an uint32 and it will work.

Secrets: Added rules for the low-level damage cap. There is a rule for the pre-10 cap, and one for the pre-20 cap. Defaults to 20 and 40 respectively.

REQUIRED: You will have to recompile sharedmem with this update. Items will fail to load otherwise.

+++ EQEmuServer/common/Item.cpp
@@ -1761,7 +1761,7 @@ void EvoItemInst::ScaleItem() {
 	m_scaledItem->BaneDmgAmt = (sint8)((float)m_item->BaneDmgAmt*Mult);
 	m_scaledItem->BardValue = (sint32)((float)m_item->BardValue*Mult);
 	m_scaledItem->ElemDmgAmt = (uint8)((float)m_item->ElemDmgAmt*Mult);
-	m_scaledItem->Damage = (uint8)((float)m_item->Damage*Mult);
+	m_scaledItem->Damage = (uint32)((float)m_item->Damage*Mult);
 
 	m_scaledItem->CombatEffects = (sint8)((float)m_item->CombatEffects*Mult);
 	m_scaledItem->Shielding = (sint8)((float)m_item->Shielding*Mult);
